requirements.txt
asgiref==3.8.1
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.8
defusedxml==0.8.0rc2
Django==5.0.6
django-environ==0.11.2
django-templated-mail==1.1.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.3
idna==3.7
oauthlib==3.2.2
pillow==10.3.0
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.5.0
urllib3==2.2.1
